WOAH! Subtle bug possibly affecting many scripts
Apparently the if the left operand of && throws an error then the exit
code will be propagated IF it is the final command in the script. The ||
operator is unaffected since only the right operand's exit code is used
(and if the right operand throws an error then you probably want that
error to be propagated). If statements are unaffected as well since they
throw the exit codes out entirely.

Most of the scripts hit by this commit probably weren't affected by this
behavior, since the affected command needs to be the final command in
the script, but it's better to be on the safe side.