plauncher: !!BREAKING!! completely changed history format
The history format and location has been changed. Switching to the new
config is simple:

mv ~/.config/plauncher/recent ~/.plauncher

Your history will work seamlessly, but plauncher will have to ask you
about your background/terminal preferences again. The advantage to this
format is that everything is now stored in a single file instead of
three, and it's trivial to move that file somewhere reasonably with the
PLAUNCHER_HISTORY variable.

The reason for the previous design was because this script was
originally a reimplementation of a Manjaro (gross) script called
dmenu_recent (or was it recency? I can't remember). I wanted to be able
to carry over the same history files, so I kept the same design. But now
we can have a more sensible design without the breakage, since the
history file is no longer managed by fzf.