stream_notify: automatically overwrite state file
Previously it was expected that the user would configure cron to delete
the state file for them. This was dumb because any streams running at
the time of deletion would get an extra notification, and it was also
just bad UI.

Now the state file just lists all of the last known live streams. This
has the disadvantage that any network outage can cause an extra
notification, but also has the advantage that it gives a notification
when a downed stream comes back up.